Revert "Snap 3269 (#197)"
This reverts commit 5a4fa893847956348ed4ed50f3fff09bad9f9fb0.

+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/state/HDFSBackedStateStoreProvider.scala
@@ -17,8 +17,7 @@
 
 package org.apache.spark.sql.execution.streaming.state
 
-import java._
-import java.io.{DataInputStream, DataOutputStream, FileNotFoundException, IOException}
+import java.io.{DataInputStream, DataOutputStream, IOException}
 
 import scala.collection.JavaConverters._
 import scala.collection.mutable
@@ -72,10 +71,9 @@ private[state] class HDFSBackedStateStoreProvider(
     valueSchema: StructType,
     storeConf: StateStoreConf,
     hadoopConf: Configuration
-) extends StateStoreProvider with Logging {
+  ) extends StateStoreProvider with Logging {
 
-  val numberOfVersionsToRetainInMemory = storeConf.maxVersionsToRetainInMemory
-  type MapType = util.HashMap[UnsafeRow, UnsafeRow]
+  type MapType = java.util.HashMap[UnsafeRow, UnsafeRow]
 
   /** Implementation of [[StateStore]] API which is backed by a HDFS-compatible file system */
   class HDFSBackedStateStore(val version: Long, mapToUpdate: MapType)
@@ -89,8 +87,9 @@ private[state] class HDFSBackedStateStoreProvider(
 
     private val newVersion = version + 1
     private val tempDeltaFile = new Path(baseDir, s"temp-${Random.nextLong}")
-    private lazy val tempDeltaFileStream = compressStream(fs.create(tempDeltaFile, true))
-    private val allUpdates = new util.HashMap[UnsafeRow, StoreUpdate]()
+    private val tempDeltaFileStream = compressStream(fs.create(tempDeltaFile, true))
+
+    private val allUpdates = new java.util.HashMap[UnsafeRow, StoreUpdate]()
 
     @volatile private var state: STATE = UPDATING
     @volatile private var finalDeltaFile: Path = null
@@ -102,7 +101,7 @@ private[state] class HDFSBackedStateStoreProvider(
     }
 
     override def put(key: UnsafeRow, value: UnsafeRow): Unit = {
-      verify(state == UPDATING, "Cannot put after already committed or aborted")
+      verify(state == UPDATING, "Cannot remove after already committed or aborted")
 
       val isNewKey = !mapToUpdate.containsKey(key)
       mapToUpdate.put(key, value)
@@ -111,7 +110,7 @@ private[state] class HDFSBackedStateStoreProvider(
         case Some(ValueAdded(_, _)) =>
           // Value did not exist in previous version and was added already, keep it marked as added
           allUpdates.put(key, ValueAdded(key, value))
-        case Some(ValueUpdated(_, _)) | Some(ValueRemoved(_, _)) =>
+        case Some(ValueUpdated(_, _)) | Some(KeyRemoved(_)) =>
           // Value existed in previous version and updated/removed, mark it as updated
           allUpdates.put(key, ValueUpdated(key, value))
         case None =>
@@ -126,25 +125,23 @@ private[state] class HDFSBackedStateStoreProvider(
     /** Remove keys that match the following condition */
     override def remove(condition: UnsafeRow => Boolean): Unit = {
       verify(state == UPDATING, "Cannot remove after already committed or aborted")
-      val entryIter = mapToUpdate.entrySet().iterator()
-      while (entryIter.hasNext) {
-        val entry = entryIter.next
-        if (condition(entry.getKey)) {
-          val value = entry.getValue
-          val key = entry.getKey
-          entryIter.remove()
+      val keyIter = mapToUpdate.keySet().iterator()
+      while (keyIter.hasNext) {
+        val key = keyIter.next
+        if (condition(key)) {
+          keyIter.remove()
 
           Option(allUpdates.get(key)) match {
             case Some(ValueUpdated(_, _)) | None =>
               // Value existed in previous version and maybe was updated, mark removed
-              allUpdates.put(key, ValueRemoved(key, value))
+              allUpdates.put(key, KeyRemoved(key))
             case Some(ValueAdded(_, _)) =>
               // Value did not exist in previous version and was added, should not appear in updates
               allUpdates.remove(key)
-            case Some(ValueRemoved(_, _)) =>
-            // Remove already in update map, no need to change
+            case Some(KeyRemoved(_)) =>
+              // Remove already in update map, no need to change
           }
-          writeToDeltaFile(tempDeltaFileStream, ValueRemoved(key, value))
+          writeToDeltaFile(tempDeltaFileStream, KeyRemoved(key))
         }
       }
     }
@@ -157,12 +154,12 @@ private[state] class HDFSBackedStateStoreProvider(
         finalizeDeltaFile(tempDeltaFileStream)
         finalDeltaFile = commitUpdates(newVersion, mapToUpdate, tempDeltaFile)
         state = COMMITTED
-        logInfo(s"Committed version $newVersion for $this to file $finalDeltaFile")
+        logInfo(s"Committed version $newVersion for $this")
         newVersion
       } catch {
         case NonFatal(e) =>
           throw new IllegalStateException(
-            s"Error committing version $newVersion into $this", e)
+            s"Error committing version $newVersion into ${HDFSBackedStateStoreProvider.this}", e)
       }
     }
 
@@ -174,10 +171,10 @@ private[state] class HDFSBackedStateStoreProvider(
       if (tempDeltaFileStream != null) {
         tempDeltaFileStream.close()
       }
-      if (tempDeltaFile != null) {
+      if (tempDeltaFile != null && fs.exists(tempDeltaFile)) {
         fs.delete(tempDeltaFile, true)
       }
-      logInfo(s"Aborted version $newVersion for $this")
+      logInfo("Aborted")
     }
 
     /**
@@ -200,17 +197,11 @@ private[state] class HDFSBackedStateStoreProvider(
       allUpdates.values().asScala.toIterator
     }
 
-    override def numKeys(): Long = mapToUpdate.size()
-
     /**
      * Whether all updates have been committed
      */
-    override private[streaming] def hasCommitted: Boolean = {
+    override private[state] def hasCommitted: Boolean = {
       state == COMMITTED
-    }
-
-    override def toString(): String = {
-      s"HDFSStateStore[id = (op=${id.operatorId}, part=${id.partitionId}), dir = $baseDir]"
     }
   }
 
@@ -222,7 +213,7 @@ private[state] class HDFSBackedStateStoreProvider(
       newMap.putAll(loadMap(version))
     }
     val store = new HDFSBackedStateStore(version, newMap)
-    logInfo(s"Retrieved version $version of ${HDFSBackedStateStoreProvider.this} for update")
+    logInfo(s"Retrieved version $version of $this for update")
     store
   }
 
@@ -238,12 +229,12 @@ private[state] class HDFSBackedStateStoreProvider(
   }
 
   override def toString(): String = {
-    s"HDFSStateStoreProvider[id = (op=${id.operatorId}, part=${id.partitionId}), dir = $baseDir]"
+    s"StateStore[id = (op=${id.operatorId}, part=${id.partitionId}), dir = $baseDir]"
   }
 
   /* Internal classes and methods */
 
-  private val loadedMaps = new util.TreeMap[Long, MapType](Ordering[Long].reverse)
+  private val loadedMaps = new mutable.HashMap[Long, MapType]
   private val baseDir =
     new Path(id.checkpointLocation, s"${id.operatorId}/${id.partitionId.toString}")
   private val fs = baseDir.getFileSystem(hadoopConf)
@@ -257,55 +248,10 @@ private[state] class HDFSBackedStateStoreProvider(
   private def commitUpdates(newVersion: Long, map: MapType, tempDeltaFile: Path): Path = {
     synchronized {
       val finalDeltaFile = deltaFile(newVersion)
-
-      // scalastyle:off
-      // Renaming a file atop an existing one fails on HDFS
-      // (http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-common/filesystem/filesystem.html).
-      // Hence we should either skip the rename step or delete the target file. Because deleting the
-      // target file will break speculation, skipping the rename step is the only choice. It's still
-      // semantically correct because Structured Streaming requires rerunning a batch should
-      // generate the same output. (SPARK-19677)
-      // scalastyle:on
-      if (fs.exists(finalDeltaFile)) {
-        fs.delete(tempDeltaFile, true)
-      } else if (!fs.rename(tempDeltaFile, finalDeltaFile)) {
-        throw new IOException(s"Failed to rename $tempDeltaFile to $finalDeltaFile")
-      }
-      putStateIntoStateCacheMap(newVersion, map)
+      fs.rename(tempDeltaFile, finalDeltaFile)
+      loadedMaps.put(newVersion, map)
       finalDeltaFile
     }
-  }
-
-  /** This method is intended to be only used for unit test(s). DO NOT TOUCH ELEMENTS IN MAP! */
-  private[state] def getLoadedMaps(): util.SortedMap[Long, MapType] = synchronized {
-    // shallow copy as a minimal guard
-    loadedMaps.clone().asInstanceOf[util.SortedMap[Long, MapType]]
-  }
-
-  private def putStateIntoStateCacheMap(newVersion: Long, map: MapType): Unit = synchronized {
-    if (numberOfVersionsToRetainInMemory <= 0) {
-      if (loadedMaps.size() > 0) loadedMaps.clear()
-      return
-    }
-
-    while (loadedMaps.size() > numberOfVersionsToRetainInMemory) {
-      loadedMaps.remove(loadedMaps.lastKey())
-    }
-
-    val size = loadedMaps.size()
-    if (size == numberOfVersionsToRetainInMemory) {
-      val versionIdForLastKey = loadedMaps.lastKey()
-      if (versionIdForLastKey > newVersion) {
-        // this is the only case which we can avoid putting, because new version will be placed to
-        // the last key and it should be evicted right away
-        return
-      } else if (versionIdForLastKey < newVersion) {
-        // this case needs removal of the last key before putting new one
-        loadedMaps.remove(versionIdForLastKey)
-      }
-    }
-
-    loadedMaps.put(newVersion, map)
   }
 
   /**
@@ -314,7 +260,7 @@ private[state] class HDFSBackedStateStoreProvider(
    */
   private[state] def latestIterator(): Iterator[(UnsafeRow, UnsafeRow)] = synchronized {
     val versionsInFiles = fetchFiles().map(_.version).toSet
-    val versionsLoaded = loadedMaps.keySet.asScala
+    val versionsLoaded = loadedMaps.keySet
     val allKnownVersions = versionsInFiles ++ versionsLoaded
     if (allKnownVersions.nonEmpty) {
       loadMap(allKnownVersions.max).entrySet().iterator().asScala.map { x =>
@@ -332,26 +278,29 @@ private[state] class HDFSBackedStateStoreProvider(
 
   /** Initialize the store provider */
   private def initialize(): Unit = {
-    try {
+    if (!fs.exists(baseDir)) {
       fs.mkdirs(baseDir)
-    } catch {
-      case e: IOException =>
+    } else {
+      if (!fs.isDirectory(baseDir)) {
         throw new IllegalStateException(
-          s"Cannot use ${id.checkpointLocation} for storing state data for $this: $e ", e)
+          s"Cannot use ${id.checkpointLocation} for storing state data for $this as " +
+            s"$baseDir already exists and is not a directory")
+      }
     }
   }
 
   /** Load the required version of the map data from the backing files */
   private def loadMap(version: Long): MapType = {
     if (version <= 0) return new MapType
-    synchronized { Option(loadedMaps.get(version)) }.getOrElse {
+    synchronized { loadedMaps.get(version) }.getOrElse {
       val mapFromFile = readSnapshotFile(version).getOrElse {
         val prevMap = loadMap(version - 1)
         val newMap = new MapType(prevMap)
+        newMap.putAll(prevMap)
         updateFromDeltaFile(version, newMap)
         newMap
       }
-      putStateIntoStateCacheMap(version, mapFromFile)
+      loadedMaps.put(version, mapFromFile)
       mapFromFile
     }
   }
@@ -379,7 +328,7 @@ private[state] class HDFSBackedStateStoreProvider(
         writeUpdate(key, value)
       case ValueUpdated(key, value) =>
         writeUpdate(key, value)
-      case ValueRemoved(key, value) =>
+      case KeyRemoved(key) =>
         writeRemove(key)
     }
   }
@@ -391,16 +340,13 @@ private[state] class HDFSBackedStateStoreProvider(
 
   private def updateFromDeltaFile(version: Long, map: MapType): Unit = {
     val fileToRead = deltaFile(version)
-    var input: DataInputStream = null
-    val sourceStream = try {
-      fs.open(fileToRead)
-    } catch {
-      case f: FileNotFoundException =>
-        throw new IllegalStateException(
-          s"Error reading delta file $fileToRead of $this: $fileToRead does not exist", f)
+    if (!fs.exists(fileToRead)) {
+      throw new IllegalStateException(
+        s"Error reading delta file $fileToRead of $this: $fileToRead does not exist")
     }
+    var input: DataInputStream = null
     try {
-      input = decompressStream(sourceStream)
+      input = decompressStream(fs.open(fileToRead))
       var eof = false
 
       while(!eof) {
@@ -459,6 +405,8 @@ private[state] class HDFSBackedStateStoreProvider(
 
   private def readSnapshotFile(version: Long): Option[MapType] = {
     val fileToRead = snapshotFile(version)
+    if (!fs.exists(fileToRead)) return None
+
     val map = new MapType()
     var input: DataInputStream = null
 
@@ -495,9 +443,6 @@ private[state] class HDFSBackedStateStoreProvider(
       }
       logInfo(s"Read snapshot file for version $version of $this from $fileToRead")
       Some(map)
-    } catch {
-      case _: FileNotFoundException =>
-        None
     } finally {
       if (input != null) input.close()
     }
@@ -512,13 +457,13 @@ private[state] class HDFSBackedStateStoreProvider(
         val lastVersion = files.last.version
         val deltaFilesForLastVersion =
           filesForVersion(files, lastVersion).filter(_.isSnapshot == false)
-        synchronized { Option(loadedMaps.get(lastVersion)) } match {
+        synchronized { loadedMaps.get(lastVersion) } match {
           case Some(map) =>
             if (deltaFilesForLastVersion.size > storeConf.minDeltasForSnapshot) {
               writeSnapshotFile(lastVersion, map)
             }
           case None =>
-          // The last map is not loaded, probably some other instance is in charge
+            // The last map is not loaded, probably some other instance is in charge
         }
 
       }
@@ -541,15 +486,13 @@ private[state] class HDFSBackedStateStoreProvider(
         if (earliestVersionToRetain > 0) {
           val earliestFileToRetain = filesForVersion(files, earliestVersionToRetain).head
           synchronized {
-            val mapsToRemove = loadedMaps.asScala.keys.filter(_ < earliestVersionToRetain).toSeq
+            val mapsToRemove = loadedMaps.keys.filter(_ < earliestVersionToRetain).toSeq
             mapsToRemove.foreach(loadedMaps.remove)
           }
-          val filesToDelete = files.filter(_.version < earliestFileToRetain.version)
-          filesToDelete.foreach { f =>
+          files.filter(_.version < earliestFileToRetain.version).foreach { f =>
             fs.delete(f.path, true)
           }
-          logInfo(s"Deleted files older than ${earliestFileToRetain.version} for $this: " +
-            filesToDelete.mkString(", "))
+          logInfo(s"Deleted files older than ${earliestFileToRetain.version} for $this")
         }
       }
     } catch {
@@ -572,7 +515,7 @@ private[state] class HDFSBackedStateStoreProvider(
 
         val deltaFiles = allFiles.filter { file =>
           file.version > snapshotFile.version && file.version <= version
-        }.toList
+        }
         verify(
           deltaFiles.size == version - snapshotFile.version,
           s"Unexpected list of delta files for version $version for $this: $deltaFiles"
@@ -590,7 +533,7 @@ private[state] class HDFSBackedStateStoreProvider(
     val files: Seq[FileStatus] = try {
       fs.listStatus(baseDir)
     } catch {
-      case _: io.FileNotFoundException =>
+      case _: java.io.FileNotFoundException =>
         Seq.empty
     }
     val versionToFiles = new mutable.HashMap[Long, StoreFile]
@@ -613,7 +556,7 @@ private[state] class HDFSBackedStateStoreProvider(
       }
     }
     val storeFiles = versionToFiles.values.toSeq.sortBy(_.version)
-    logDebug(s"Current set of files for $this: ${storeFiles.mkString(", ")}")
+    logDebug(s"Current set of files for $this: $storeFiles")
     storeFiles
   }
